Refactor Shard struct: remove dead code and optimize layout
Major changes:
- Remove unused function pointers: mutate, crossover, getState, setState, resetState (40 bytes saved)
- Replace SHBool owned with uint32_t flags bitfield for future extensibility
- Change id from uint64_t to uint32_t (sufficient for shard identification)
- Remove unused debuggerId field
- Optimize field layout for cache locality:
  * Move inlineShardId to offset 0 (accessed FIRST in VM loop)
  * Pack all hot fields in first 64-byte cache line:
    - VM execution: inlineShardId, activate
    - Reference counting: refCount
    - Profiling: name, nameLength
    - Error logging: line, column, file
    - Debug: id, flags
    - Common functions: hash, warmup

The getState/setState were only used in serialization but stored redundant data that's already reconstructed from parameters during warmup. All five removed function pointers had zero or redundant implementations across the codebase.

Updated all call sites to use bitwise flag operations (SHARD_FLAGS_OWNED_SHARD) instead of direct boolean access.

Fixed serialization to use correct sizeof(uint32_t) for id field instead of sizeof(uint64_t) which was corrupting the flags field during deserialization.

Final struct size: 200 bytes (down from 240+ bytes)
All frequently accessed fields fit perfectly in first cache line (64 bytes)

+++ shards/lang/src/eval.rs
@@ -2652,7 +2652,7 @@ fn shard_with_id_ex(shard: AutoShardRef, e: &mut EvalEnv, x: DebugPtr) -> AutoSh
     id
   });
   if let Some(id) = id {
-    unsafe { (*shard.0 .0).id = id };
+    unsafe { (*shard.0 .0).id = id as u32 };
   }
   shard
 }
